requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers==4.46.3
pillow
pandas
matplotlib
seaborn
numpy
sentencepiece
protobuf
accelerate